requirements.txt
streamlit
google-generativeai
pymongo
requests
python-dotenv
pandas
dnspython
jira
pydantic
jinja2
fastapi
uvicorn
fastapi-middleware
botbuilder-core
botbuilder-schema
aiohttp